post_classify_image/requirements.txt
tensorflow>=2.11
pillow
h5py
boto3 == 1.26.5
ImageHash